Fixed unused var issues
Next removed as an argument for route handlers in index.js as it was unnecessary.

waitToScrape now an async function with a sleep function that returns a new Promise wrapping the setTimeout, which resolves after the delay. This is awaited before running the recursive getStoriesData call, which is returned after it is awaited.