Theme data persists in local storage
themeContextProvider now receives theme state from a sticky state hook that checks localStorage for a key named userPreferredTheme, and returns it if it exists, otherwise returns the default value.

In _app.js, the context provider is wrapped in a stateful component called ClientOnly that forces a rerender after the initial render by setting the hasMounted state to true. This means that when the context provider calls the useStickyState hook, the useStickyState will have access to the window.localStorage.

Note: This totally defeats the purpose of using Next.js as nothing is being pre-rendered due to the ClientOnly component wrapping all components, but it was good to learn about client-side-only rendering and about using local storage.