Scraper only scrapes new content
Scrapes won't be inserted to the db if they already exist.

The getNewestStoryId function is called at the start of runCron and the return value is passed down to scrapePage, where it's compared to the id of the element being scraped.

mongoUtil object built and exported from the db instantiation module (mongoUtil) to enable the reuse of the same connection across the app.