Frontend gets stories in batches
Frontend requests 3 stories at initial render and then increments the new batchRequest state variable which is initialised at zero. This variable is appended to the fetch request string as a query string.

The server multiplies this value by the skipMultiplier so subsequent requests skip over the documents that the frontend has already received and fetch another batch of the same size.