Frontend retrieves more stories as needed
When the unreadStories array gets down to 5 elements, a request is made to the server to top it up.

The new isMounted state var is set to true on initial render, and an Effect sets the currentStory after it is true.

Removed the completedStories state variable and the functionality to move stories from unreadStories to it. It was unnecessary now that new stories can be pulled in from the db.

newStory() renamed updateUnreadStories, because without the completeStories variable, that's all that that function does now anyway.