No more documents in db error fixed
A fetch call is made to /count at initial render to get the number of docs in the db. This is saved to state.

After fetching more docs from the db, a check is completed to see if there are enough docs to fill the next batch. If not, the batchRequest state var is reset to 0 so the next request will pick documents from the start of the db.